Inverted FocusEnvironment
Updated examples/environment.ipynb.
Had Dynamics use the new common TypeVar.
Added a more flexible base type for environmental actions.
Simplified dynamics_test with the new generic actions.
Created episode_ender and episode_ender_test.
Inverted control of FocusEnvironment.
Split the dependencies of FocusEnvironment into a number of object.
Split focus_environment_test into a number of more appropriate tests.
Created focus_instrument and focus_instrument_test.
Created observation_normer and observation_normer_test.
Created observation_producer and observation_producer_test.
Created observation_rewarder and observation_rewarder_test.
Created state_initializer and state_initializer_test.
Collected common environmental types into one file.
Added an enum type for safe(r) observation and state access.
Added a fancy visualizer for performance history!
Added render.make_render_target to make frame initialization simpler.
Switched rendering to uin8 from some mishmash of that and float32.
Simplified graphics/world.

+++ reinfocus/vision.py
@@ -18,5 +18,5 @@ def focus_value(image: NDArray) -> float:
         a better focus."""
 
     return cv2.Laplacian(
-        cv2.medianBlur(cv2.cvtColor(image, cv2.COLOR_RGB2GRAY), 3), cv2.CV_32F
+        cv2.medianBlur(cv2.cvtColor(image, cv2.COLOR_RGB2GRAY), 3), cv2.CV_8U
     ).var()
